Change outlier rejection to a quorum-based approach
Previous method allowed for 2-point spikes to pass as real data.
Changed such that it is only considered a real measurement if it is within the threshold for at least N points. If it is not within the threshold for at least that many points, it is an outlier.

Only jumps that are more than N points long will be counted, and they will only be counted starting at the N+1'th point.
To be stricter, increase the quorum amount, decrease the reading_keep_amount (never below the quorum amount), and decrease the threshold.